+++ lib/capistrano/logtail/utility.rb
@@ -10,7 +10,12 @@ module Capistrano
       def tail(*files)
         @context.info("Tailing #{files.join(' ')}")
 
-        cmd = "tail -f #{files.join(' ')}"
+        cmd = [
+            :tail,
+            '-f',
+            "-n#{lines}",
+            *files
+          ].join(' ')
 
         @context.send(:with_ssh) do |ssh|
           ssh.open_channel do |chan|
@@ -23,6 +28,20 @@ module Capistrano
           end
           ssh.loop
         end
+      end
+
+      # Rails environment
+      #
+      # @return [String] the rails environment.
+      def rails_env
+        fetch(:rails_env, fetch(:stage, 'production')).to_s
+      end
+
+      # Number of log lines
+      #
+      # @return [Integer] number of existing log lines to show.
+      def lines
+        (ENV['n'] || fetch(:logtail_lines, 10)).to_i
       end
     end
   end