+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
 ## Unreleased (initial release)
-###Features
+### Features
 - Tailing of all logs, rails log and custom log file
+- Allow to specify default log files and number of existing lines to show